Rename to pkg/servicerepo (from common/servicerepo).
