+++ services/domains/domains.go
@@ -8,7 +8,7 @@ import (
 	"github.com/trustwallet/blockatlas/pkg/blockatlas"
 	"github.com/trustwallet/blockatlas/pkg/errors"
 	"github.com/trustwallet/blockatlas/platform"
-	"github.com/trustwallet/blockatlas/common/servicerepo"
+	"github.com/trustwallet/blockatlas/pkg/servicerepo"
 )
 
 type DomainsServiceI interface {
